+++ RELEASE_NOTES.md
@@ -67,5 +67,8 @@ found 0 vulnerabilities
 
 Developed by **Maximus Prime**
 
-- GitHub: [@MaximusPrime77](https://github.com/MaximusPrime77)
+- GitHub: [@MaximusPrime](https://github.com/MaximusPrime)
+- Organization: [@MaximusPrimeSoftware](https://github.com/MaximusPrimeSoftware)
+- Website: [maximusprimesoftware.pages.dev](https://maximusprimesoftware.pages.dev/)
 - Email: [b.maximus.prime@gmail.com](mailto:b.maximus.prime@gmail.com)
+- Studio email: [maximusprimesoftware@gmail.com](mailto:maximusprimesoftware@gmail.com)